JWT authentication added in api and route protected

+++ routes/api/items.js
@@ -1,38 +1,54 @@
 const express = require("express");
 const router = express.Router();
+const auth = require("../../middlewares/auth");
 
 //Item Model
 const ItemModel = require("../../models/Items")
 
 //End Points
-router.get("/", async (req, res) => {
+router.get("/", auth, async (req, res) => {
     try {
-        const Items = await ItemModel.find();
-        res.json(Items);
+        const AllItems = await ItemModel.find();
+        const Items = [];
+        AllItems.forEach(note => {
+            if (note.author == req.user.id) {
+                Items.push(note);
+            }
+        })
+        res.status(200).json(Items);
     } catch (err) {
-        res.json(err);
+        res.status(400).json(err);
     }
 })
 
-router.post("/", async (req, res) => {
+router.post("/", auth, async (req, res) => {
     try {
         const newItem = new ItemModel({
             title: req.body.title,
-            content: req.body.content
+            content: req.body.content,
+            author: req.user.id
         });
         const Items = await newItem.save();
-        res.json(Items);
+        res.status(201).json(Items);
     } catch (err) {
-        res.json(err);
+        res.status(400).json(err);
     }
 })
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", auth, async (req, res) => {
     try {
-        const Item = await ItemModel.deleteOne({ _id: req.params.id });
-        res.json(Item);
+        let Item = await ItemModel.findById(req.params.id);
+        if (Item.author == req.user.id) {
+            const DeletedItem = await ItemModel.deleteOne({ _id: req.params.id });
+            res.status(200).json(DeletedItem);
+        }
+        else {
+            res.status(401).json({
+                msg: "You can not delete this item"
+            });
+        }
     } catch (err) {
-        res.json(err);
+        res.status(400).json(err);
     }
 })
